수정

+++ studyroom/cron.py
@@ -1,6 +1,9 @@
 from .models import Reserve
 from datetime import datetime, timedelta
 
+
 def schedule_week():
-    q = Reserve.objects.all()
-    q.delete()
+    real_time = datetime.datetime.now().weekday()
+    if real_time == 6:
+        q = Reserve.objects.all()
+        q.delete()
